backend/requirements.txt
fastapi[all]>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
supabase>=2.5.0,<3.0.0
pydantic[email]>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
reportlab>=4.2.0,<5.0.0
python-multipart>=0.0.9,<1.0.0
slowapi>=0.1.9,<0.2.0
structlog>=24.0.0,<25.0.0
bandit>=1.8.0,<2.0.0
stripe>=11.0.0,<12.0.0
